% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{get_color_list}
\alias{get_color_list}
\title{Generate a List of Colors}
\usage{
get_color_list(n)
}
\arguments{
\item{n}{Integer. The number of colors to return. Must be 18 or smaller.}
}
\value{
A character vector of hexadecimal color codes.
}
\description{
This function returns a specified number of colors from a predefined palette.
}
\details{
The function selects colors based on a predefined index order and
returns the first \code{n} colors in the sorted order.
}
\examples{
# Get the first 5 colors
get_color_list(5)

# Get all 18 colors
get_color_list(18)

}
